+++ src/test/java/com/mattsmeets/macrokey/handler/GameTickHandlerTest.java
@@ -3,11 +3,12 @@ package com.mattsmeets.macrokey.handler;
 import com.mattsmeets.macrokey.event.ExecuteOnTickEvent;
 import com.mattsmeets.macrokey.event.InGameTickEvent;
 import com.mattsmeets.macrokey.event.MacroActivationEvent;
-import com.mattsmeets.macrokey.model.command.CommandInterface;
 import com.mattsmeets.macrokey.model.Macro;
 import com.mattsmeets.macrokey.model.MacroInterface;
+import com.mattsmeets.macrokey.model.command.CommandInterface;
 import com.mattsmeets.macrokey.model.lambda.ExecuteOnTickInterface;
 import net.minecraft.client.entity.EntityPlayerSP;
+import net.minecraft.client.settings.KeyBinding;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
@@ -29,12 +30,13 @@ public class GameTickHandlerTest {
         MacroActivationEvent event = mock(MacroActivationEvent.class);
 
         Macro macro = mock(Macro.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
         when(event.getMacroState()).thenReturn(MacroActivationEvent.MacroState.KEY_DOWN);
         Set<MacroInterface> macros = Collections.singleton(macro);
         when(event.getMacros()).thenReturn(macros);
 
-        GameTickHandler handler = new GameTickHandler(spySet, null);
+        GameTickHandler handler = new GameTickHandler(spySet, null, new KeyBinding[]{keyBinding});
 
         handler.onKeyEvent(event);
 
@@ -51,10 +53,11 @@ public class GameTickHandlerTest {
         MacroActivationEvent event = mock(MacroActivationEvent.class);
 
         Macro macro = mock(Macro.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
         when(event.getMacroState()).thenReturn(MacroActivationEvent.MacroState.KEY_UP);
 
-        GameTickHandler handler = new GameTickHandler(spySet, null);
+        GameTickHandler handler = new GameTickHandler(spySet, null, new KeyBinding[]{keyBinding});
 
         handler.onKeyEvent(event);
 
@@ -70,8 +73,9 @@ public class GameTickHandlerTest {
         ExecuteOnTickEvent event = mock(ExecuteOnTickEvent.class);
 
         Macro macro = mock(Macro.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
-        GameTickHandler handler = new GameTickHandler(null, spySet);
+        GameTickHandler handler = new GameTickHandler(null, spySet, new KeyBinding[]{keyBinding});
 
         ExecuteOnTickInterface executor = (boolean delay) -> {
         };
@@ -95,8 +99,9 @@ public class GameTickHandlerTest {
         InGameTickEvent event = mock(InGameTickEvent.class);
         Macro macro = mock(Macro.class);
         ExecuteOnTickInterface executor = mock(ExecuteOnTickInterface.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
-        GameTickHandler handler = new GameTickHandler(macroSet, executorSet);
+        GameTickHandler handler = new GameTickHandler(macroSet, executorSet, new KeyBinding[]{keyBinding});
 
         handler.onTick(event);
 
@@ -114,8 +119,9 @@ public class GameTickHandlerTest {
         InGameTickEvent event = mock(InGameTickEvent.class);
         Macro macro = mock(Macro.class);
         ExecuteOnTickInterface executor = mock(ExecuteOnTickInterface.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
-        GameTickHandler handler = new GameTickHandler(macroSet, executorSet);
+        GameTickHandler handler = new GameTickHandler(macroSet, executorSet, new KeyBinding[]{keyBinding});
 
         handler.onExecutorEvent(new ExecuteOnTickEvent(executor));
 
@@ -138,8 +144,9 @@ public class GameTickHandlerTest {
         InGameTickEvent event = mock(InGameTickEvent.class);
         Macro macro = mock(Macro.class);
         ExecuteOnTickInterface executor = mock(ExecuteOnTickInterface.class);
+        KeyBinding keyBinding = mock(KeyBinding.class);
 
-        GameTickHandler handler = new GameTickHandler(macroSet, executorSet);
+        GameTickHandler handler = new GameTickHandler(macroSet, executorSet, new KeyBinding[]{keyBinding});
 
         when(event.isLimitedTick()).thenReturn(true);
 
@@ -184,7 +191,9 @@ public class GameTickHandlerTest {
         inputMacros.add(macro1);
 
         MacroActivationEvent macroActivationEvent = mock(MacroActivationEvent.class);
-        GameTickHandler handler = new GameTickHandler(macroSet, null);
+        KeyBinding keyBinding = mock(KeyBinding.class);
+
+        GameTickHandler handler = new GameTickHandler(macroSet, null, new KeyBinding[]{keyBinding});
 
         when(macroActivationEvent.getMacroState()).thenReturn(MacroActivationEvent.MacroState.KEY_DOWN);
         when(macroActivationEvent.getMacros()).thenReturn(inputMacros);
@@ -232,7 +241,9 @@ public class GameTickHandlerTest {
         inputMacros.add(macro1);
 
         MacroActivationEvent macroActivationEvent = mock(MacroActivationEvent.class);
-        GameTickHandler handler = new GameTickHandler(macroSet, null);
+        KeyBinding keyBinding = mock(KeyBinding.class);
+
+        GameTickHandler handler = new GameTickHandler(macroSet, null, new KeyBinding[]{keyBinding});
 
         when(macroActivationEvent.getMacroState()).thenReturn(MacroActivationEvent.MacroState.KEY_DOWN);
         when(macroActivationEvent.getMacros()).thenReturn(inputMacros);
